ext4: fix DAX write locking
Unlike O_DIRECT DAX is not an optional opt-in feature selected by the
application, so we'll have to provide the traditional synchronіzation
of overlapping writes as we do for buffered writes.

This was broken historically for DAX, but got fixed for ext2 and XFS
as part of the iomap conversion.  Fix up ext4 as well.

Reviewed-by: Jan Kara <jack@suse.cz>
Change-Id: 8f7ae49865e38ad026f8959daf769b16a8d8d12e
Signed-off-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Signed-off-by: TogoFire <togofire@mailfence.com>